# apis/api/v1/ci_task/ci_task.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.5
// 	protoc        v5.29.3
// source: v1/ci_task/ci_task.proto

package task

import (
	_ "github.com/envoyproxy/protoc-gen-validate/validate"
	_ "google.golang.org/genproto/googleapis/api/annotations"
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
	app "kubecaptain/apis/api/v1/app"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

type CreateRequest struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	AppName       *app.NameRequest       `protobuf:"bytes,1,opt,name=app_name,json=appName,proto3" json:"app_name,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *CreateRequest) Reset() {
	*x = CreateRequest{}
	mi := &file_v1_ci_task_ci_task_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *CreateRequest) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*CreateRequest) ProtoMessage() {}

func (x *CreateRequest) ProtoReflect() protoreflect.Message {
	mi := &file_v1_ci_task_ci_task_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use CreateRequest.ProtoReflect.Descriptor instead.
func (*CreateRequest) Descriptor() ([]byte, []int) {
	return file_v1_ci_task_ci_task_proto_rawDescGZIP(), []int{0}
}

func (x *CreateRequest) GetAppName() *app.NameRequest {
	if x != nil {
		return x.AppName
	}
	return nil
}

var File_v1_ci_task_ci_task_proto protoreflect.FileDescriptor

var file_v1_ci_task_ci_task_proto_rawDesc = string([]byte{
	0x0a, 0x18, 0x76, 0x31, 0x2f, 0x63, 0x69, 0x5f, 0x74, 0x61, 0x73, 0x6b, 0x2f, 0x63, 0x69, 0x5f,
	0x74, 0x61, 0x73, 0x6b, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x1b, 0x6b, 0x75, 0x62, 0x65,
	0x63, 0x61, 0x70, 0x74, 0x61, 0x69, 0x6e, 0x2e, 0x61, 0x70, 0x69, 0x73, 0x2e, 0x61, 0x70, 0x69,
	0x2e, 0x76, 0x31, 0x2e, 0x61, 0x70, 0x70, 0x1a, 0x1c, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f,
	0x61, 0x70, 0x69, 0x2f, 0x61, 0x6e, 0x6e, 0x6f, 0x74, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x73, 0x2e,
	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x1f, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x61, 0x70,
	0x69, 0x2f, 0x66, 0x69, 0x65, 0x6c, 0x64, 0x5f, 0x62, 0x65, 0x68, 0x61, 0x76, 0x69, 0x6f, 0x72,
	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x1b, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70,
	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x65, 0x6d, 0x70, 0x74, 0x79, 0x2e, 0x70, 0x72,
	0x6f, 0x74, 0x6f, 0x1a, 0x17, 0x76, 0x61, 0x6c, 0x69, 0x64, 0x61, 0x74, 0x65, 0x2f, 0x76, 0x61,
	0x6c, 0x69, 0x64, 0x61, 0x74, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x1a, 0x10, 0x76, 0x31,
	0x2f, 0x61, 0x70, 0x70, 0x2f, 0x61, 0x70, 0x70, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x5a,
	0x0a, 0x0d, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x12,
	0x49, 0x0a, 0x08, 0x61, 0x70, 0x70, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28,
	0x0b, 0x32, 0x28, 0x2e, 0x6b, 0x75, 0x62, 0x65, 0x63, 0x61, 0x70, 0x74, 0x61, 0x69, 0x6e, 0x2e,
	0x61, 0x70, 0x69, 0x73, 0x2e, 0x61, 0x70, 0x69, 0x2e, 0x76, 0x31, 0x2e, 0x61, 0x70, 0x70, 0x2e,
	0x4e, 0x61, 0x6d, 0x65, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x42, 0x04, 0xe2, 0x41, 0x01,
	0x02, 0x52, 0x07, 0x61, 0x70, 0x70, 0x4e, 0x61, 0x6d, 0x65, 0x32, 0x8c, 0x01, 0x0a, 0x10, 0x41,
	0x70, 0x70, 0x43, 0x49, 0x54, 0x61, 0x73, 0x6b, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x12,
	0x78, 0x0a, 0x06, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x12, 0x2a, 0x2e, 0x6b, 0x75, 0x62, 0x65,
	0x63, 0x61, 0x70, 0x74, 0x61, 0x69, 0x6e, 0x2e, 0x61, 0x70, 0x69, 0x73, 0x2e, 0x61, 0x70, 0x69,
	0x2e, 0x76, 0x31, 0x2e, 0x61, 0x70, 0x70, 0x2e, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x52, 0x65,
	0x71, 0x75, 0x65, 0x73, 0x74, 0x1a, 0x16, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70,
	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x45, 0x6d, 0x70, 0x74, 0x79, 0x22, 0x2a, 0x82,
	0xd3, 0xe4, 0x93, 0x02, 0x24, 0x3a, 0x01, 0x2a, 0x22, 0x1f, 0x2f, 0x76, 0x31, 0x2f, 0x61, 0x70,
	0x70, 0x2f, 0x7b, 0x61, 0x70, 0x70, 0x5f, 0x6e, 0x61, 0x6d, 0x65, 0x2e, 0x6e, 0x61, 0x6d, 0x65,
	0x7d, 0x2f, 0x63, 0x69, 0x2f, 0x74, 0x61, 0x73, 0x6b, 0x42, 0x2a, 0x5a, 0x28, 0x6b, 0x75, 0x62,
	0x65, 0x63, 0x61, 0x70, 0x74, 0x61, 0x69, 0x6e, 0x2f, 0x61, 0x70, 0x69, 0x73, 0x2f, 0x61, 0x70,
	0x69, 0x2f, 0x76, 0x31, 0x2f, 0x61, 0x70, 0x70, 0x2f, 0x63, 0x69, 0x2f, 0x74, 0x61, 0x73, 0x6b,
	0x3b, 0x74, 0x61, 0x73, 0x6b, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
})

var (
	file_v1_ci_task_ci_task_proto_rawDescOnce sync.Once
	file_v1_ci_task_ci_task_proto_rawDescData []byte
)

func file_v1_ci_task_ci_task_proto_rawDescGZIP() []byte {
	file_v1_ci_task_ci_task_proto_rawDescOnce.Do(func() {
		file_v1_ci_task_ci_task_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_v1_ci_task_ci_task_proto_rawDesc), len(file_v1_ci_task_ci_task_proto_rawDesc)))
	})
	return file_v1_ci_task_ci_task_proto_rawDescData
}

var file_v1_ci_task_ci_task_proto_msgTypes = make([]protoimpl.MessageInfo, 1)
var file_v1_ci_task_ci_task_proto_goTypes = []any{
	(*CreateRequest)(nil),   // 0: kubecaptain.apis.api.v1.app.CreateRequest
	(*app.NameRequest)(nil), // 1: kubecaptain.apis.api.v1.app.NameRequest
	(*emptypb.Empty)(nil),   // 2: google.protobuf.Empty
}
var file_v1_ci_task_ci_task_proto_depIdxs = []int32{
	1, // 0: kubecaptain.apis.api.v1.app.CreateRequest.app_name:type_name -> kubecaptain.apis.api.v1.app.NameRequest
	0, // 1: kubecaptain.apis.api.v1.app.AppCITaskService.Create:input_type -> kubecaptain.apis.api.v1.app.CreateRequest
	2, // 2: kubecaptain.apis.api.v1.app.AppCITaskService.Create:output_type -> google.protobuf.Empty
	2, // [2:3] is the sub-list for method output_type
	1, // [1:2] is the sub-list for method input_type
	1, // [1:1] is the sub-list for extension type_name
	1, // [1:1] is the sub-list for extension extendee
	0, // [0:1] is the sub-list for field type_name
}

func init() { file_v1_ci_task_ci_task_proto_init() }
func file_v1_ci_task_ci_task_proto_init() {
	if File_v1_ci_task_ci_task_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_v1_ci_task_ci_task_proto_rawDesc), len(file_v1_ci_task_ci_task_proto_rawDesc)),
			NumEnums:      0,
			NumMessages:   1,
			NumExtensions: 0,
			NumServices:   1,
		},
		GoTypes:           file_v1_ci_task_ci_task_proto_goTypes,
		DependencyIndexes: file_v1_ci_task_ci_task_proto_depIdxs,
		MessageInfos:      file_v1_ci_task_ci_task_proto_msgTypes,
	}.Build()
	File_v1_ci_task_ci_task_proto = out.File
	file_v1_ci_task_ci_task_proto_goTypes = nil
	file_v1_ci_task_ci_task_proto_depIdxs = nil
}