# apis/api/v1/ci_task/ci_task_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.8.4
// - protoc             v5.29.3
// source: v1/ci_task/ci_task.proto

package task

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationAppCITaskServiceCreate = "/kubecaptain.apis.api.v1.app.AppCITaskService/Create"

type AppCITaskServiceHTTPServer interface {
	Create(context.Context, *CreateRequest) (*emptypb.Empty, error)
}

func RegisterAppCITaskServiceHTTPServer(s *http.Server, srv AppCITaskServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/app/{app_name.name}/ci/task", _AppCITaskService_Create1_HTTP_Handler(srv))
}

func _AppCITaskService_Create1_HTTP_Handler(srv AppCITaskServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAppCITaskServiceCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*CreateRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type AppCITaskServiceHTTPClient interface {
	Create(ctx context.Context, req *CreateRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type AppCITaskServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewAppCITaskServiceHTTPClient(client *http.Client) AppCITaskServiceHTTPClient {
	return &AppCITaskServiceHTTPClientImpl{client}
}

func (c *AppCITaskServiceHTTPClientImpl) Create(ctx context.Context, in *CreateRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/v1/app/{app_name.name}/ci/task"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAppCITaskServiceCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}